requirements.txt
discord.py>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncpg>=0.28.0
typing_extensions>=4.0.0